[Rust] Temporarily introduce an "initrd2" while migrating to Rust initrd

+++ programs/subprojects/libutils/sleep.c
@@ -9,7 +9,7 @@
 #define AXLE_CORE_SERVICE_NAME "com.axle.core"
 #define AMC_SLEEP_UNTIL_TIMESTAMP 202
 
-unsigned usleep(unsigned int ms) {
+int usleep(long unsigned int ms) {
     uint32_t b[2];
     b[0] = AMC_SLEEP_UNTIL_TIMESTAMP;
     b[1] = ms;
